task 6

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 	char charType;
 	int intType;
 	int longintType;
-	int doubleType;
+	double doubleType;
 	float floatType;
 
 	/* sizeof evaluates the size if variable */
